Reworked scan-build
The new diagnostic command is now "analyze-build, which uses clang's
analyze-build instead of scan-build. The main reason for this change
was that cmake sometimes didn't use the right compiler when invoked
with scan-build. With analyze-build there are no problems like this
as analyze-build uses only compile commands JSON file, which is
generated by cmake, but it doesn't need cmake invocation to do the
analysis.